Update safe stock logic

+++ models/area_sales.server.model.js
@@ -19,8 +19,8 @@ module.exports = function (appDb) {
     department: {
       type: String
     },
-    is_sure:{
-      type:String
+    is_sure: {
+      type: String
     },
     //上月月份
     month: {
@@ -184,7 +184,7 @@ module.exports = function (appDb) {
       this.next_month_sales_forecast_5 = parseInt((this.next_month_sales_forecast_2 + this.next_month_sales_forecast_3 + this.next_month_sales_forecast_4 ) / 3);
       this.next_month_sales_forecast_6 = parseInt((this.next_month_sales_forecast_3 + this.next_month_sales_forecast_4 + this.next_month_sales_forecast_5 ) / 3);
 
-      var area_safe_stock = isNaN(parseInt(this.area_safe_stock)) ? 0 : parseInt(this.area_safe_stock);
+      var area_safe_stock = isNaN(parseFloat(this.area_safe_stock)) ? 0 : parseFloat(this.area_safe_stock);
       this.safe_stock = parseInt(( this.last_month_sales_count_3 + this.last_month_sales_count_2 + this.last_month_sales_count_1) / 3) * area_safe_stock;
     }
