+++ models/hq_sales.server.model.js
@@ -31,8 +31,8 @@ module.exports = function (appDb) {
     hq_safe_stock: {
       type: String,
     },
-    is_sure:{
-      type:String
+    is_sure: {
+      type: String
     },
     month: {
       type: String
@@ -221,7 +221,7 @@ module.exports = function (appDb) {
       this.next_month_sales_forecast_4 = parseInt((this.next_month_sales_forecast_1 + this.next_month_sales_forecast_2 + this.next_month_sales_forecast_3) / 3);
       this.next_month_sales_forecast_5 = parseInt((this.next_month_sales_forecast_2 + this.next_month_sales_forecast_3 + this.next_month_sales_forecast_4 ) / 3);
       this.next_month_sales_forecast_6 = parseInt((this.next_month_sales_forecast_3 + this.next_month_sales_forecast_4 + this.next_month_sales_forecast_5 ) / 3);
-      var hq_safe_stock = isNaN(parseInt(this.hq_safe_stock)) ? 0 : parseInt(this.hq_safe_stock);
+      var hq_safe_stock = isNaN(parseFloat(this.hq_safe_stock)) ? 0 : parseFloat(this.hq_safe_stock);
       this.safe_stock = parseInt(( this.last_month_sales_count_3 + this.last_month_sales_count_2 + this.last_month_sales_count_1) / 3) * hq_safe_stock;
     }
